VLDB/requirements.txt
neo4j
python-dotenv
sqlalchemy
Faker
ollama
impyla
networkx
matplotlib